# control-plane/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Control-plane is a central microservice of Service Mesh which is responsible for all the gateways configuration. For more information, visit our Documentation(https://github.com/Netcracker/qubership-core-control-plane/tree/main/README.md).",
        "title": "Control Plane API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v3/composite-platform/namespaces": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Composite Structure",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Get Composite Structure",
                "operationId": "HandleGetCompositeStructure",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/composite.Structure"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v3/composite-platform/namespaces/{namespace}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Namespace To Composite",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Add Namespace To Composite",
                "operationId": "AddNamespaceToComposite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove Namespace From Composite",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Remove Namespace From Composite",
                "operationId": "RemoveNamespaceFromComposite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "composite.Structure": {
            "type": "object",
            "properties": {
                "baseline": {
                    "type": "string"
                },
                "satellites": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}